Restructured gulp tasks into separate files
