+++ package.json
@@ -106,6 +106,7 @@
     "webpack": "^1.13.1",
     "webpack-dev-server": "^1.14.1",
     "webpack-merge": "^0.8.4",
+    "wrench": "^1.5.9",
     "yargs": "^4.7.1"
   }
 }
